Add files via upload
Added 1-4 link unique armor search. Added a label in the unique armor search to indicate that the result would be 1-4 link armor values only. Added a button in the unique weapon search to display the highest valued weapon currently on the path of exile market. Still adjusting and messing around with label and button placements.