Add files via upload
Added support for searching fossils, Added a button the main menu for fossils, as well as fossil search button in the fossil window, as well as a button to display the most valuable fossil on the market. implemented code to give titles to the Tkinter windows, respective to the window the user is in.