Add files via upload
Added search functionality for unique Accessories. Unique Accessories in path of exile are unique rings, amulets, and belts. Added a button in the unique Accessories search window to display the most valuable accessory. Started moving the position of main menu buttons in order to make them more aesthetic, This is a work in progress.