Add files via upload
Adjusted the layout of the main menu, I might still change this a little more in the future, Added search functionality for divination cards, as well as a button to display the most expensive div card on the market. Added a new button that allows the user to enter a float value, and calculates the ratio of exalts to chaos orbs at the current market price.